requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
aiosmtplib==2.0.2
aiostream==0.5.2
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
bcrypt==4.0.1
beautifulsoup4==4.12.2
bleach==6.1.0
blinker==1.7.0
boto3==1.34.1
botocore==1.34.1
Brotli==1.1.0
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.17
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
cryptography==41.0.5
databases==0.8.0
dataclasses-json==0.5.14
Deprecated==1.2.14
deprecation==2.1.0
distro==1.8.0
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.1.0.post1
exceptiongroup==1.1.3
fastapi==0.104.1
fastapi-mail==1.4.1
filelock==3.13.1
flatbuffers==23.5.26
frozenlist==1.4.0
fsspec==2023.10.0
google-api-core==2.14.0
google-auth==2.23.4
googleapis-common-protos==1.61.0
greenlet==3.0.1
grpcio==1.59.3
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.1
huggingface-hub==0.19.4
humanfriendly==10.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.1
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
kubernetes==28.1.0
langchain==0.0.337
langsmith==0.0.65
llama-index==0.9.3.post1
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-enum==1.5.1
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.8
networkx==3.2.1
nltk==3.8.1
numexpr==2.8.7
numpy==1.26.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
onnxruntime==1.16.2
openai==1.3.7
openapi-schema-pydantic==1.2.4
opentelemetry-api==1.21.0
opentelemetry-exporter-otlp-proto-common==1.21.0
opentelemetry-exporter-otlp-proto-grpc==1.21.0
opentelemetry-proto==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-semantic-conventions==0.42b0
orjson==3.9.10
overrides==7.4.0
packaging==23.2
pandas==2.1.3
passlib==1.7.4
Pillow==10.1.0
pip==23.3.1
pluggy==1.3.0
posthog==3.0.2
protobuf==4.24.4
psycopg2-binary==2.9.9
pulsar-client==3.3.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.1
pydantic_core==2.14.3
pydantic-extra-types==2.1.0
pydantic-settings==2.1.0
pyOpenSSL==23.3.0
PyPika==0.48.9
PySocks==1.7.1
pytest==7.4.3
pytest-subtests==0.11.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
python-slugify==8.0.1
pytz==2023.3.post1
pyu2f==0.1.5
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rsa==4.9
s3transfer==0.9.0
safetensors==0.4.0
scikit-learn==1.3.2
scipy==1.11.4
sentence-transformers==2.2.2
sentencepiece==0.1.99
setuptools==68.2.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==1.4.50
sqlalchemy-redshift==0.8.14
starlette==0.27.0
stringcase==1.2.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
text-unidecode==1.3
threadpoolctl==3.2.0
tiktoken==0.5.1
tokenizers==0.15.0
tomli==2.0.1
torch==2.1.1
torchvision==0.16.1
tqdm==4.66.1
transformers==4.35.2
triton==2.1.0
typer==0.9.0
typing_extensions==4.8.0
typing-inspect==0.9.0
tzdata==2023.3
ujson==5.8.0
urllib3==1.26.18
uvicorn==0.24.0.post1
uvloop==0.19.0
watchfiles==0.21.0
webencodings==0.5.1
websocket-client==1.6.4
websockets==12.0
wheel==0.41.3
wrapt==1.16.0
yarl==1.9.2
zipp==3.17.0
python-socketio==5.11.0
bidict==0.22.1
build==1.0.3
dotenv==0.0.5
langchain-community==0.0.13
langchain-core==0.1.12
mmh3==4.1.0
opentelemetry-instrumentation==0.43b0
opentelemetry-instrumentation-asgi==0.43b0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-util-http==0.43b0
psycopg2==2.9.9
pyproject_hooks==1.0.0
pyreadline3==3.4.1
python-engineio==4.8.2
simple-websocket==1.0.0
wsproto==1.2.0